[CLOUD-2542] fixing need of shared storage in the algorithm in general

+++ os-eap-txnrecovery/added/partitionPV.sh
@@ -1,16 +1,25 @@
 #!/bin/sh
+# -------
+# Handling start of application server with transaction recovery migration pod
+#   it manages the safety of the transactions when object store can be touched
+#   by application server and the migration pod
+# -------
 
 source $JBOSS_HOME/bin/launch/launch-common.sh
+source $JBOSS_HOME/bin/launch/logging.sh
 
 [ "x${SCRIPT_DEBUG}" = "xtrue" ] && DEBUG_QUERY_API_PARAM="-l debug"
 
-# when jdbc object store used for transaction data we can save recovery marker to database too
-# TX_JDBC_RECOVERY_MARKER_PERMITTED: can be forced by user
-TX_JDBC_RECOVERY_MARKER_PERMITTED=${TX_JDBC_RECOVERY_MARKER_PERMITTED:-true}
-# IS_TX_SQL_BACKEND: depends on definition of jdbc credentials to connect to database
+# data should be placed to shared partition
+IS_SPLIT_DATA_DEFINED=false
+[ "${SPLIT_DATA^^}" = "TRUE" ] && IS_SPLIT_DATA_DEFINED=true
+# saving recovery marker data to database
 IS_TX_SQL_BACKEND=false
-if [ -n "${TX_DATABASE_PREFIX_MAPPING}" ] && [ "x$TX_JDBC_RECOVERY_MARKER_PERMITTED" = "xtrue" ]; then
-    IS_TX_SQL_BACKEND=true
+[ -n "${TX_DATABASE_PREFIX_MAPPING}" ] && IS_TX_SQL_BACKEND=true
+[ "x$FORBID_TX_JDBC_RECOVERY_MARKER" = "xtrue" ] && IS_TX_SQL_BACKEND=false
+if ! $IS_SPLIT_DATA_DEFINED && [ "x$FORBID_TX_JDBC_RECOVERY_MARKER" = "xtrue" ]; then
+    log_warning "[`date`] Forbidden to use jdbc for recovery marker by use of variable FORBID_TX_JDBC_RECOVERY_MARKER while SPLIT_DATA is not enabled."
+    log_warning "The transaction recovery marker algorithm is not capable to save recovery data and won't work."
 fi
 
 # parameters
@@ -33,10 +42,9 @@ function init_pod_name() {
   # openshift sets up the node id as host name
   [ -z "${POD_NAME}" ] && POD_NAME="${HOSTNAME}"
 
-  # having set the POD_NAME is crucial as the processing depends on unique
-  # pod name being used as identifier for migration
+  # having set the POD_NAME is crucial as the migration process depends on unique identifier
   if [ -z "${POD_NAME}" ]; then
-    >&2 echo "Cannot proceed further as failed to get unique POD_NAME as identifier of the server to be started"
+    log_error "[`date`] Cannot proceed further as failed to get unique POD_NAME as identifier of the server to be started"
     exit 1
   fi
 }
@@ -54,42 +62,48 @@ function truncate_jboss_node_name() {
 
 # parameters
 # - base directory
-function partitionPV() {
+function startApplicationServer() {
   local podsDir="$1"
-  local applicationPodDir
-
-  mkdir -p "${podsDir}"
-
+  # 1) pods is available to start the application server
   init_pod_name
+
   local applicationPodDir="${podsDir}/${POD_NAME}"
-
   $IS_TX_SQL_BACKEND && initJdbcRecoveryMarkerProperties
+  $IS_SPLIT_DATA_DEFINED && mkdir -p "${podsDir}"
 
+  # 2) while any recovery marker matches, sleep and wait for recovery to finish
   local waitCounter=0
-  # 2) while any file matching, sleep
   while true; do
 
     if isRecoveryInProgress; then
-      echo "Waiting to start pod ${POD_NAME} as recovery process '$(echo ${podsDir}/${POD_NAME}-RECOVERY-*)' is currently cleaning data directory."
+      log_info "[`date`] Waiting to start pod ${POD_NAME} as recovery process for the pod is currently in progress"
     else
       # no recovery running: we are free to start the app container
       break
     fi
-
     sleep 1
-    echo "`date`: waiting for recovery process to clean the environment for the pod to start"
   done
 
-  # 3) create /pods/<applicationPodName>
-  SERVER_DATA_DIR="${applicationPodDir}/serverData"
-  mkdir -p "${SERVER_DATA_DIR}"
+  # 3) create app server data directory with name of the pod name /or/ creating pod dir jdbc record
+  if $IS_TX_SQL_BACKEND; then
+    SERVER_DATA_DIR="${podsDir}"
 
-  if [ ! -f "${SERVER_DATA_DIR}/../data_initialized" ]; then
-    init_data_dir ${SERVER_DATA_DIR}
-    touch "${SERVER_DATA_DIR}/../data_initialized"
+    local applicationPodExistence=($(${JDBC_COMMAND_PODNAME_REGISTRY} select_application -a "${applicationPodDir}"))
+    if [ "x${applicationPodExistence}" = "x" ]; then
+      ${JDBC_COMMAND_PODNAME_REGISTRY} insert -a "${applicationPodDir}" -r 'undefined'
+    fi
+  fi
+  if $IS_SPLIT_DATA_DEFINED; then
+    SERVER_DATA_DIR="${applicationPodDir}/serverData"
+    mkdir -p "${SERVER_DATA_DIR}"
+
+    if [ ! -f "${SERVER_DATA_DIR}/../data_initialized" ]; then
+      init_data_dir ${SERVER_DATA_DIR}
+      touch "${SERVER_DATA_DIR}/../data_initialized"
+    fi
   fi
 
-  # 4) launch server with node name as pod name (openshift-node-name.sh uses the node name value)
+  # 4) launch server with NODE_NAME to be assigned as POD_NAME (openshift-node-name.sh uses it)
   NODE_NAME=$(truncate_jboss_node_name "${POD_NAME}") runServer "${SERVER_DATA_DIR}" &
 
   PID=$!
@@ -101,10 +115,10 @@ function partitionPV() {
   trap - TERM
   wait $PID 2>/dev/null
 
-  echo "Server terminated with status $STATUS ($(kill -l $STATUS 2>/dev/null))"
+  log_info "[`date`] Server terminated with status $STATUS ($(kill -l $STATUS 2>/dev/null))"
 
   if [ "$STATUS" -eq 255 ] ; then
-    echo "Server returned 255, changing to 254"
+    log_info "[`date`] Server returned 255, changing to 254"
     STATUS=254
   fi
 
@@ -127,10 +141,20 @@ function migratePV() {
 
   while true ; do
 
-    # 1) Periodically, for each /pods/<applicationPodName>
-    for applicationPodDir in "${podsDir}"/*; do
-      # check if the found file is type of directory, if not directory move to the next item
-      [ ! -d "$applicationPodDir" ] && continue
+    if $IS_TX_SQL_BACKEND; then
+      applicationPodDirs=($(${JDBC_COMMAND_PODNAME_REGISTRY} select_application))
+    else
+      unset applicationPodDirs
+      declare -a applicationPodDirs
+      for potentialApplicationPodDir in "${podsDir}"/*; do
+          # check if the found file is type of directory, if not directory move to the next item
+          [ ! -d "$potentialApplicationPodDir" ] && continue
+          applicationPodDirs+=("$potentialApplicationPodDir")
+      done
+    fi
+
+    # 1) Periodically, for each directory /pods/<applicationPodName> /or/ record in jdbc registry
+    for applicationPodDir in "${applicationPodDirs[@]}"; do
 
       # 1.a) create the recovery marker
       local applicationPodName="$(basename ${applicationPodDir})"
@@ -138,10 +162,10 @@ function migratePV() {
       STATUS=42 # expecting there could be  error on getting living pods
 
       # 1.a.i) if <applicationPodName> is not in the cluster
-      echo "examining existence of living pod for directory: '${applicationPodDir}'"
+      log_info "[`date`] Examining existence of living pod '${applicationPodName}'"
       unset LIVING_PODS
       LIVING_PODS=($($(dirname ${BASH_SOURCE[0]})/queryosapi.py -q pods_living -f list_space ${DEBUG_QUERY_API_PARAM}))
-      [ $? -ne 0 ] && echo "ERROR: Can't get list of living pods" && continue
+      [ $? -ne 0 ] && log_warning "[`date`] Can't get list of living pods" && continue
       # expecting the application pod of the same name was started/is living, it will manage recovery on its own
       local IS_APPLICATION_POD_LIVING=true
       if ! arrContains ${applicationPodName} "${LIVING_PODS[@]}"; then
@@ -163,10 +187,10 @@ function migratePV() {
           trap - TERM
           wait $PID 2>/dev/null
 
-          echo "Migration terminated with status $STATUS ($(kill -l $STATUS))"
+          log_info "[`date`] Migration of pod ${applicationPodName} terminated with status $STATUS ($(kill -l $STATUS))"
 
           if [ "$STATUS" -eq 255 ] ; then
-            echo "Server returned 255, changing to 254"
+            log_info "[`date`] Server returned 255, changing to 254"
             STATUS=254
           fi
           exit $STATUS
@@ -182,9 +206,19 @@ function migratePV() {
         wait $PID 2>/dev/null
 
         if [ $STATUS -eq 0 ]; then
-          # 1.a.iii) Delete /pods/<applicationPodName> when recovery was succesful
-          echo "`date`: Migration succesfully finished for application directory ${applicationPodDir} thus removing it by recovery pod ${recoveryPodName}"
-          rm -rf "${applicationPodDir}"
+          # 1.a.iii) Delete /pods/<applicationPodName> directory /or/ jdbc registry record when recovery was succesful
+          log_info "[`date`] Migration finished for pod name ${applicationPodName}, the record removed by recovery pod ${recoveryPodName}"
+          if $IS_TX_SQL_BACKEND; then
+            ${JDBC_COMMAND_PODNAME_REGISTRY} delete -a ${applicationPodDir}
+
+            # removing not only recovery marker record but the narayana object store table too
+            local podprefix="os${applicationPodName//-/}"
+            local narayanasuffix="JBossTSTxTable"; # narayana does not exposes this value
+            log_info "Dropping Narayana object store table '${podprefix}${narayanasuffix}'"
+            ${JDBC_COMMAND_PODNAME_REGISTRY} drop -t "${podprefix}${narayanasuffix}"
+          else
+            rm -rf "${applicationPodDir}"
+          fi
         fi
       fi
 
@@ -199,7 +233,7 @@ function migratePV() {
       recoveryPodsGarbageCollection
     done
 
-    echo "`date`: Finished Migration Check cycle, pausing for ${MIGRATION_PAUSE} seconds before resuming"
+    log_info "[`date`] Finished migration check cycle, pausing for ${MIGRATION_PAUSE} seconds before resuming"
     MIGRATION_POD_TIMESTAMP=$(getPodLogTimestamp)
     trap 'kill $(jobs -p)' EXIT
     sleep "${MIGRATION_PAUSE}" & wait
@@ -213,10 +247,11 @@ function isRecoveryInProgress() {
   local isRecoveryInProgress=false
   if $IS_TX_SQL_BACKEND; then
     # jdbc based recovery descriptor
-    recoveryMarkers=($(${JDBC_RECOVERY_MARKER_COMMAND} select_recovery -a ${POD_NAME}))
+    recoveryMarkers=($(${JDBC_COMMAND_RECOVERY_MARKER} select_recovery -a ${POD_NAME}))
     local isRecoveryInProgress=false
     [ ${#recoveryMarkers[@]} -ne 0 ] && isRecoveryInProgress=true # array is not empty, there are recovery markers existing
-  else
+  fi
+  if $IS_SPLIT_DATA_DEFINED; then
     # shared file system based recovery descriptor
     find "${podsDir}" -maxdepth 1 -type f -name "${POD_NAME}-RECOVERY-*" 2>/dev/null | grep -q .
     # is there an existing RECOVERY descriptor that means a recovery is in progress
@@ -236,7 +271,7 @@ function createRecoveryMarker() {
 
   if $IS_TX_SQL_BACKEND; then
     # jdbc recovery marker insertion
-    ${JDBC_RECOVERY_MARKER_COMMAND} insert -a ${applicationPodName} -r ${recoveryPodName}
+    ${JDBC_COMMAND_RECOVERY_MARKER} insert -a ${applicationPodName} -r ${recoveryPodName}
   else
     # file system recovery marker creation: /pods/<applicationPodName>-RECOVERY-<recoveryPodName>
     touch "${podsDir}/${applicationPodName}-RECOVERY-${recoveryPodName}"
@@ -255,7 +290,7 @@ function removeRecoveryMarker() {
 
   if $IS_TX_SQL_BACKEND; then
     # jdbc recovery marker removal
-    ${JDBC_RECOVERY_MARKER_COMMAND} delete -a ${applicationPodName} -r ${recoveryPodName}
+    ${JDBC_COMMAND_RECOVERY_MARKER} delete -a ${applicationPodName} -r ${recoveryPodName}
   else
     # file system recovery marker deletion
     rm -f "${podsDir}/${applicationPodName}-RECOVERY-${recoveryPodName}"
@@ -268,17 +303,17 @@ function removeRecoveryMarker() {
 function recoveryPodsGarbageCollection() {
   local livingPods=($($(dirname ${BASH_SOURCE[0]})/queryosapi.py -q pods_living -f list_space ${DEBUG_QUERY_API_PARAM}))
   if [ $? -ne 0 ]; then # fail to connect to openshift api
-    echo "ERROR: Can't get list of living pods. Can't do recovery marker garbage collection."
+    log_warning "[`date`] Can't get list of living pods. Can't do recovery marker garbage collection."
     return 1
   fi
 
   if $IS_TX_SQL_BACKEND; then
     # jdbc
-    local recoveryMarkers=($(${JDBC_RECOVERY_MARKER_COMMAND} select_recovery))
+    local recoveryMarkers=($(${JDBC_COMMAND_RECOVERY_MARKER} select_recovery))
     for recoveryPod in ${recoveryMarkers[@]}; do
       if ! arrContains ${recoveryPod} "${livingPods[@]}"; then
         # recovery pod is dead, garbage collecting
-        ${JDBC_RECOVERY_MARKER_COMMAND} delete -r ${recoveryPod}
+        ${JDBC_COMMAND_RECOVERY_MARKER} delete -r ${recoveryPod}
       fi
     done
   else
@@ -320,7 +355,7 @@ function probePodLogForRecoveryErrors() {
   local probeStatus=$?
 
   if [ $probeStatus -ne 0 ]; then
-    echo "Cannot contact OpenShift API to get log for pod ${POD_NAME}"
+    log_warning "[`date`] Cannot contact OpenShift API to get log for pod ${podNameToProbe}"
     return 1
   fi
 
@@ -334,7 +369,7 @@ function probePodLogForRecoveryErrors() {
   [ "x${SCRIPT_DEBUG}" = "xtrue" ] && set -x
 
   if $isPeriodicRecoveryError; then # ERROR string was found in the log output
-    echo "Pod '${POD_NAME}' started with periodic recovery errors: '$line'"
+    log_warning "[`date`] Pod '${podNameToProbe}' started with periodic recovery errors: '$line'"
     return 1
   fi
 
@@ -358,7 +393,11 @@ function initJdbcRecoveryMarkerProperties() {
   JDBC_RECOVERY_USER=$(find_env "${prefix}_USERNAME")
   JDBC_RECOVERY_PASSWORD=$(find_env "${prefix}_PASSWORD")
   JDBC_RECOVERY_NAMESPACE=$(cat /var/run/secrets/kubernetes.io/serviceaccount/namespace)
-  JDBC_RECOVERY_TABLE="recmark_${TX_JDBC_RECOVERY_MARKER_TABLE_SUFFIX//-/}"
+
+  local jdbcTableSuffix=${TX_JDBC_RECOVERY_MARKER_TABLE_SUFFIX//-/}
+  [ "x${jdbcTableSuffix}" != "x" ] && jdbcTableSuffix="_${jdbcTableSuffix}"
+  JDBC_RECOVERY_TABLE="recmark${jdbcTableSuffix}"
+  JDBC_PODNAME_REGISTRY_TABLE="recpodreg${jdbcTableSuffix}"
 
     case "$db" in
     "MYSQL")
@@ -373,7 +412,12 @@ function initJdbcRecoveryMarkerProperties() {
   # do not reduce logging when debug is enabled
   [ "x${SCRIPT_DEBUG}" = "xtrue" ] && LOGGING_PROPERTIES=""
 
-  JDBC_RECOVERY_MARKER_COMMAND="java $LOGGING_PROPERTIES -jar $JBOSS_HOME/jboss-modules.jar -mp $JBOSS_HOME/modules/ io.narayana.openshift-recovery -y ${JDBC_RECOVERY_DB_TYPE} -o ${JDBC_RECOVERY_DB_HOST} -p ${JDBC_RECOVERY_DB_PORT} -d ${JDBC_RECOVERY_DATABASE} -u ${JDBC_RECOVERY_USER} -s ${JDBC_RECOVERY_PASSWORD} -t ${JDBC_RECOVERY_TABLE} -c"
+  # one table works as storage for recovery markers, other stores information about started pods
+  local jdbcCommand="java $LOGGING_PROPERTIES -jar $JBOSS_HOME/jboss-modules.jar -mp $JBOSS_HOME/modules/ io.narayana.openshift-recovery -y ${JDBC_RECOVERY_DB_TYPE} -o ${JDBC_RECOVERY_DB_HOST} -p ${JDBC_RECOVERY_DB_PORT} -d ${JDBC_RECOVERY_DATABASE} -u ${JDBC_RECOVERY_USER} -s ${JDBC_RECOVERY_PASSWORD}"
+  JDBC_COMMAND_RECOVERY_MARKER="${jdbcCommand} -t ${JDBC_RECOVERY_TABLE} -c"
+  JDBC_COMMAND_PODNAME_REGISTRY="${jdbcCommand} -t ${JDBC_PODNAME_REGISTRY_TABLE} -c"
+
   # creating the database schema
-  ${JDBC_RECOVERY_MARKER_COMMAND} create
+  ${JDBC_COMMAND_RECOVERY_MARKER} create
+  ${JDBC_COMMAND_PODNAME_REGISTRY} create
 }
